make script replayable

+++ init.sh
@@ -3,12 +3,15 @@
 export DEBIAN_FRONTEND="noninteractive"
 
 apt-get update && apt-get dist-upgrade -y 
-curl -s https://packages.cloud.google.com/apt/doc/apt-key.gpg | apt-key add -
-echo "deb http://apt.kubernetes.io/ kubernetes-xenial main" >> /etc/apt/sources.list.d/kubernetes.list 
-apt-get update && apt-get install -y docker.io kubelet kubeadm
 
-systemctl stop kubelet
-rm -rf /var/lib/kubelet
+if [ ! -f /etc/apt/sources.list.d/kubernetes.list ] ; then
+    curl -s https://packages.cloud.google.com/apt/doc/apt-key.gpg | apt-key add -
+    echo "deb http://apt.kubernetes.io/ kubernetes-xenial main" >> /etc/apt/sources.list.d/kubernetes.list 
+    apt-get update && apt-get install -y docker.io kubelet kubeadm
+
+    systemctl stop kubelet
+    rm -rf /var/lib/kubelet
+fi
 
 sudo -u dil mkdir -p /home/dil/.ssh
 [ ! -f /home/dil/.ssh/id_ecdsa ] && sudo -u dil ssh-keygen -t ecdsa -b 521 -f /home/dil/.ssh/id_ecdsa -N ''
@@ -19,15 +22,16 @@ echo 'APT::Periodic::Unattended-Upgrade "1";' >> /etc/apt/apt.conf.d/10periodic
 
 adduser dil docker
 
-#mkdir -p /tmp/scripts
-#cd /tmp/scripts
-#git clone https://github.com/dirk1492/azure-init-script.git
+if [ -d /tmp/scripts/azure-init-script ]; then
+    git pull
+else
+    mkdir -p /tmp/scripts
+    cd /tmp/scripts
+    git clone https://github.com/dirk1492/azure-init-script.git
+fi
+
 echo "$1" >> /tmp/token.txt
 
-echo "pwd: $(pwd)"
-
-env
-
 if [ -n "$2" ]; then
-    sh "$2"
+    sh "/tmp/scripts/azure-init-script/$2"
 fi
